new router: scanner /move_to /scan /reboot /shutdown /get_scanner
got moved from main to scanner
